Добавил проверки на пустоту коллекций

+++ Replication/Replication.Core/Actors/DeleteDataObjectsActor.cs
@@ -41,10 +41,12 @@ namespace NuClear.Replication.Core.Actors
             var changes = _changesProvider.GetChanges(commandsToExecute);
 
             var toDelete = changes.Complement.ToArray();
-
-            events.AddRange(_dataChangesHandler.HandleRelates(toDelete));
-            events.AddRange(_dataChangesHandler.HandleDeletes(toDelete));
-            _bulkRepository.Delete(toDelete);
+            if (toDelete.Length != 0)
+            {
+                events.AddRange(_dataChangesHandler.HandleRelates(toDelete));
+                events.AddRange(_dataChangesHandler.HandleDeletes(toDelete));
+                _bulkRepository.Delete(toDelete);
+            }
 
             return events;
         }
